requirements.txt
agno
streamlit
pandas
openai
google-generativeai
duckdb
openpyxl
google-genai
